Cartons to one decimal, dd/mm/yyyy timestamps, Edit Scope radios, modal width
Cartons carry one decimal place (migration 030). Repositories/services round on
write and the pair count still has to land whole, since stock is counted in
indivisible pairs; the entry forms say so up front via cartonsProblem() instead
of letting the user discover it on save, and every displayed total goes through
formatCartons() so float noise never reaches the screen.

Timestamps render dd/mm/yyyy HH:mm through the new formatDateTime(). The default
locale could not be trusted: Electron's --lang=en-GB moves Chromium's UI locale
but leaves ICU resolving to en-US, so toLocaleString() was printing American
dates next to British date pickers.

Product Setup: a DUPLICATE_NAME conflict now carries the existing article_id and
name, so entering a name that already exists adds a COLOUR to that article —
which is the normal reason the same name is typed twice — rather than dead-ending.

Edit Scope: the Master/Detail picker is two bare radios (EditScopeRadios) across
all eight voucher pages. The old card was parked in the card's left gutter behind
a 2xl media query and vanished entirely at 90% zoom, taking the radios with it;
the gutter variant is absolute beside the toolbar's New button behind no query at
all, and the Receipts/Payments variant keeps its column's old width so the centre
card does not move.

ConfirmModal: dropped .card-white from the panel. That class declares
max-width:100% after Tailwind's utilities in index.css, so it beat max-w-md by
source order and the dialog stretched the full viewport width.

+++ backend/src/repositories/stock.repository.js
@@ -34,7 +34,7 @@ async function insertMovement(movement) {
       movementType: { type: sql.VarChar(15), value: movement.movement_type },
       qtyPairs: { type: sql.Int, value: movement.qty_pairs },
       movementDate: { type: sql.Date, value: movement.movement_date },
-      inputQty: { type: sql.Int, value: movement.input_qty ?? null },
+      inputQty: { type: sql.Decimal(12, 1), value: movement.input_qty ?? null },
       inputUnit: { type: sql.VarChar(10), value: movement.input_unit ?? null },
       packing: { type: sql.Int, value: movement.packing ?? null },
       createdBy: { type: sql.Int, value: movement.created_by ?? null },
